requirements.txt
pydantic-settings
python-dotenv
python-telegram-bot
requests
selenium
types-requests
